Extract sentinel callbacks
